Added PatchData::get_subpatch() function.  Here's the description and signature:
 //! Fills a PatchData with the elements attached to a center vertex.
 //! Note that all entities in the sub-patch are copies of the entities
 //! in 'this' patch.  As such, moving a vertex in the sub-patch
 //! won't move the corresponding vertex in the source patch.  Also,
 //! calling 'update_mesh()' on the sub-patch WILL modify the TSTT
 //! mesh, but the source patch won't see the changes.

void get_subpatch(size_t center_vertex_index,
                  PatchData &pd_to_fill,
                  MsqError &err);

+++ src/Mesh/MeshSet.cpp
@@ -372,7 +372,7 @@ bool MeshSet::get_next_patch(PatchData &pd, PatchDataParameters &pd_params, MsqE
                                    &tstt_err);
       assert(!tstt_err);
         // ... enters the coordinates of those vertices in PatchData ...
-      int vtx_ind[MAX_NUM_VERTICES_PER_ELEM];
+      size_t vtx_ind[MAX_NUM_VERTICES_PER_ELEM];
       for (int n=0; n<num_element_vtx; ++n)
       {
         vtx_ind[n] = pd.add_vertex(currentVertex->mesh, vertices[n],
@@ -528,7 +528,7 @@ bool MeshSet::get_next_patch(PatchData &pd, PatchDataParameters &pd_params, MsqE
          assert(!tstt_err);
 
          // ... enters the coordinates of those vertices in PatchData ...
-         int vtx_ind[MAX_NUM_VERTICES_PER_ELEM];
+         size_t vtx_ind[MAX_NUM_VERTICES_PER_ELEM];
          for (int n=0; n<num_element_vtx; ++n) {
             // retrieves the fixed tag value
             int* on_boundary = NULL;
@@ -550,27 +550,27 @@ bool MeshSet::get_next_patch(PatchData &pd, PatchDataParameters &pd_params, MsqE
 
          // ... and adds the element to PatchData.
          pd.add_element(*currentMesh, current_mesh_elements[e],
-                           vtx_ind, tstt_to_mesquite(element_topologies[e]), err);
+                        vtx_ind, tstt_to_mesquite(element_topologies[e]), err);
          MSQ_CHKERR(err);
       }
       
       delete[] element_topologies;
       delete[] elem_vtx;
       delete[] elem_vtx_coords;
-  
-      //cout << "numVertices for local patch: " << pd.num_vertices() << endl;
-  
-      // TODO : provide patch for several layers of adjacencies.
-  
-      // free entities allocated in TSTT::Entity_GetAdjacencies()
+      
+        //cout << "numVertices for local patch: " << pd.num_vertices() << endl;
+      
+        // TODO : provide patch for several layers of adjacencies.
+      
+        // free entities allocated in TSTT::Entity_GetAdjacencies()
       TSTT::Mesh_FreeEntityHandles(*currentMesh, current_mesh_elements, &tstt_err);
     }
-  
+    
     return true;
   }  
-
+  
   
   err.set_msg("no implementation for specified patch type.");
   return false;
-
+  
 }
